reorganizing group code
