+++ src/recipes/group-recipe.ts
@@ -1,9 +1,32 @@
-import { ResolverType, Iterate, isOrderAssured } from '../types-internal';
+import { isOrderAssured } from '../types-internal';
 import { AnyMapper } from '../types-internal';
 import { AnyIterable } from 'augmentative-iterable';
-import { map } from '../sync/map';
-import { orderedGroupRecipe } from './ordered-group-recipe';
+import { map as mapSync } from '../sync/map';
 import { GroupIngredients } from './ingredients';
+import { orderedOperationRecipe } from './ordered-operation-recipe';
+
+function orderedGroupRecipe({
+  map,
+  resolver,
+  partition,
+  forEach,
+}: GroupIngredients) {
+  const ordered = orderedOperationRecipe(map, resolver, partition);
+  return function <T>(this: AnyIterable<T>, mapper: AnyMapper<any>) {
+    const partitioned = ordered.call(this, mapper);
+    return map.call(partitioned, (part: Iterable<any[]>) => {
+      let key!: any;
+      const values: any[] = [];
+      return resolver(
+        forEach.call(part, ([k, v]: any) => {
+          key = k;
+          values.push(v);
+        }),
+        () => ({ key, values }),
+      );
+    });
+  };
+}
 
 export function groupRecipe(ingredients: GroupIngredients) {
   const orderedGroup = orderedGroupRecipe(ingredients);
@@ -26,7 +49,10 @@ export function groupRecipe(ingredients: GroupIngredients) {
       );
 
       const resolved = resolver(reduced, (r) =>
-        map.call(r, (([key, values]: [any, any]) => ({ key, values })) as any),
+        mapSync.call(r, (([key, values]: [any, any]) => ({
+          key,
+          values,
+        })) as any),
       );
       return iterate(resolved);
     }
